+++ CHANGELOG.md
@@ -11,9 +11,14 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 ### Changed
 * Asset management rewrite (pull request [#244]).
 * Use RON as config format ([#269])
+* Overhaul input system ([#247]), ([#261]), and ([#274])
 
 [#244]: https://github.com/amethyst/amethyst/pull/244
+[#247]: https://github.com/amethyst/amethyst/pull/247
+[#261]: https://github.com/amethyst/amethyst/pull/261
 [#269]: https://github.com/amethyst/amethyst/pull/269
+[#274]: https://github.com/amethyst/amethyst/pull/274
+
 
 ## [0.4.3] - 2017-06-03
 ### Added